ai-service-main/requirements.txt
torch
torchvision
Pillow
scikit-learn
kafka-python
requests
opencv-python-headless
